input_worker/requirements.txt
python-dotenv
pika

